Type mismatches give compiler warnings. Fixed.
git-svn-id: svn://rootdirectory.ddns.net/pdclib/branches/shepherd@635 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ functions/stdio/_PDCLIB_print.c
@@ -614,7 +614,7 @@ static int testprintf( char * buffer, const char * format, ... )
     status.write = testcb;
     va_start( status.arg, format );
     memset( buffer, '\0', 100 );
-    if ( _PDCLIB_print( format, &status ) != strlen( format ) )
+    if ( _PDCLIB_print( format, &status ) != (int)strlen( format ) )
     {
         printf( "_PDCLIB_print() did not return end-of-specifier on '%s'.\n", format );
         ++TEST_RESULTS;
